secured bank account services for logged in users

+++ service/bank_account_service.py
@@ -1,12 +1,10 @@
 from collections.abc import Sequence
 
-from exceptions.bank_account_exceptions import (
-    AccountNotFoundError,
-    DuplicateAccountNumberError,
-)
+from exceptions.bank_account_exceptions import AccountNotFoundError, DuplicateAccountNumberError, ForbiddenError
 from exceptions.user_service_exceptions import (
     UserNotFoundError,
 )
+from models.user import User
 from repositories.bank_account_repository import (
     BankAccount,
     BankAccountRepository,
@@ -23,16 +21,16 @@ class BankAccountService:
         self.user_repo = user_repo
         self.bank_repo = bank_repo
 
-    def create_account(self, account_data: BankAccountCreate) -> BankAccount:
-        user = self.user_repo.find_by_id(account_data.user_id)
+    def create_account(self, account_data: BankAccountCreate, user_id: int) -> BankAccount:
+        user = self.user_repo.find_by_id(user_id)
         if user is None:
-            raise UserNotFoundError(f"User '{account_data.user_id}' not found")
+            raise UserNotFoundError(f"User '{user_id}' not found")
         if self.bank_repo.exists_by_account_number(account_data.account_number):
             raise DuplicateAccountNumberError(
                 f"Account Number '{account_data.account_number}' already exists"
             )
         bank_account = BankAccount(
-            user_id=account_data.user_id,
+            user_id=user_id,
             bank_name=account_data.bank_name,
             account_name=account_data.account_name,
             account_number=account_data.account_number,
@@ -41,10 +39,12 @@ class BankAccountService:
         )
         return self.bank_repo.create(bank_account)
 
-    def get_account_by_id(self, account_id: int) -> BankAccount:
+    def get_account_by_id(self, account_id: int, current_user: User,) -> BankAccount:
         account = self.bank_repo.find_by_id(account_id)
         if account is None:
             raise AccountNotFoundError(f"Account '{account_id}' not found")
+        if account.user_id != current_user.id:
+            raise ForbiddenError("You are not authorized to access this account.")
         return account
 
     def get_accounts_by_user(self, user_id: int) -> Sequence[BankAccount]:
@@ -53,13 +53,23 @@ class BankAccountService:
             raise UserNotFoundError(f"User '{user_id}' not found")
         return self.bank_repo.find_all_by_user(user_id)
 
-    def update_account_name(self, account_id: int, new_name) -> BankAccount:
+    def update_account_name(self, account_id: int, current_user: User, new_name) -> BankAccount:
+        account = self.bank_repo.find_by_id(account_id)
+        if account is None:
+            raise AccountNotFoundError(f"Account '{account_id}' not found")
+        if account.user_id != current_user.id:
+            raise ForbiddenError("You are not authorized to access this account.")
         updated_account = self.bank_repo.update_account_name(account_id, new_name)
         if updated_account is None:
             raise AccountNotFoundError(f"Account '{account_id}' not found")
         return updated_account
 
-    def delete_account(self, account_id: int) -> bool:
+    def delete_account(self, account_id: int, current_user: User) -> bool:
+        account = self.bank_repo.find_by_id(account_id)
+        if account is None:
+            raise AccountNotFoundError(f"Account '{account_id}' not found")
+        if account.user_id != current_user.id:
+            raise ForbiddenError("You are not authorized to access this account.")
         deleted_account = self.bank_repo.delete_account(account_id)
         if not deleted_account:
             raise AccountNotFoundError(f"Account '{account_id}' not found")
